# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden: <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Antworten auf gültige <code>POST</code>- und <code>PATCH</code>-Anfragen beinhalten nichts.</b></li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
servers:
...
paths:
  /nutzer:
    get:
      tags:
        - Anwender
      summary: Gibt Nutzer zurück.
      description: >-
        Alle **Nutzer** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: geschlecht
          in: query
          description: 'Nur Nutzer, deren **Geschlecht** `geschlecht` ist, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        nutzerid:
                          type: integer
                          description: '**RowID des Nutzers**.'
                      required:
                        - nutzerid
                    - $ref: '#/components/schemas/nutzer'
    post:
      tags:
        - Anwender
      summary: Fügt einen Nutzer hinzu.
      description: >-
        Ein **Nutzer** wird mit den entsprechenden Werten hinzugefügt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/nutzer'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
...
  /kommentare:
    get:
      tags:
        - Authentifizierter und als Nutzer authorisierter Anwender
      summary: Gibt (eigene) Kommentare zurück.
      security:
        - basicAuth: []
      description: >-
        Alle **Kommentare** des Nutzers werden zurückgegeben. Falls Query-Parameter mit Werten angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: text
          in: query
          description: 'Nur Kommentare, deren **Text** `text` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        kommentarid:
                          type: integer
                          description: '**RowID des Kommentars**.'
                        nutzerid:
                          type: integer
                          description: '**RowID des Nutzers**.'
                        blogeintragid:
                          type: integer
                          description: '**RowID des Blogeintrags**.'
                      required:
                        - kommentarid
                        - nutzerid
                        - blogeintragid
                    - $ref: '#/components/schemas/kommentar'
  /kommentare/{kommentarid}:
    patch:
      tags:
        - Authentifizierter und als Nutzer authorisierter Anwender
      security:
        - basicAuth: []
      summary: Ändert einen (eigenen) Kommentar.
      description: >-
        Der **Kommentar** wird gemäß der Query-Parameter geändert.
      parameters:
        - name: kommentarid
          in: path
          description: '**RowID** des Kommentars.'
          schema:
            type: integer
          required: true
        - name: text
          in: query
          description: '**Text** des Kommentars.'
          schema:
            type: string
          required: true
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /blogeintraege/{blogeintragid}/schlagworte:
    post:
      tags:
        - Authentifizierter und als Redakteur authorisierter Anwender
      summary: Fügt ein Schlagwort zu einem Blogeintrag hinzu.
      description: >-
        Ein **Schlagwort** wird zu einem Blogeintrag hinzugefügt.
      security:
        - basicAuth: []
      parameters:
        - name: blogeintragid
          in: path
          description: '**RowID** des Blogeintrags.'
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/schlagwort'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    get:
      tags:
        - Anwender
      summary: Gibt die Schlagworte eines Blogeintrags zurück.
      description: >-
        Gibt alle **Schlagworte** eines Blogeintrags zurück.
      parameters:
        - name: blogeintragid
          in: path
          description: '**RowID** des Blogeintrags.'
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        schlagwortid:
                          type: integer
                          description: '**RowID des Schlagworts**.'
                      required:
                        - schlagwortid
                    - $ref: '#/components/schemas/schlagwort'
        '404':
          $ref: '#/components/responses/404'
...
  /blogeintraege/{blogeintragid}/kommentare:
    post:
      tags:
        - Authentifizierter und als Nutzer authorisierter Anwender
      summary: Fügt einen Kommentar zu einem Blogeintrag hinzu.
      description: >-
        Ein **Kommentar** wird einem Blogeintrag hinzugefügt. Das **Erstellungsdatum** wird  mit dem aktuellen Zeitstempel belegt.
      security:
        - basicAuth: []
      parameters:
        - name: blogeintragid
          in: path
          description: '**RowID** des Blogeintrags.'
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                    - type: object
                      properties:
                        text:
                          type: string
                          description: '**Text des Kommentars**.'
                      required:
                        - text
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    get:
      tags:
        - Anwender
      summary: Gibt Kommentare zu einem Blogeintrag zurück.
      description: >-
        Gibt alle **Kommentare** zu einem Blogeintrag zurück.
      parameters:
        - name: blogeintragid
          in: path
          description: '**RowID** des Blogeintrags.'
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        nutzerid:
                          type: integer
                          description: '**RowID des Nutzers**.'
                        blogeintragid:
                          type: integer
                          description: '**RowID des Blogeintrags**.'
                      required:
                        - nutzerid
                        - blogeintragid
                    - $ref: '#/components/schemas/kommentar'
        '404':
          $ref: '#/components/responses/404'
...
  '/blogeintraege/{blogeintragid}':
    delete:
      tags:
        - Authentifizierter und als Chefredakteur authorisierter Anwender
      security:
        - basicAuth: []
      summary: Löscht einen Blogeintrag.
      description: Ein **Blogeintrag** wird entfernt. Ggf. existierende **referenzierende Tupel** werden ebenfalls gelöscht.
      parameters:
        - name: blogeintragid
          in: path
          description: '**RowID** des Blogeintrages.'
          schema:
            type: integer
          required: true
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'        
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource. Siehe hierzu auch die Anmerkung bzgl. fiktiver Links im ersten Abschnitt.
          schema:
            type: string
    '204':
      description: '**No Content**.'
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
    '404':
      description: '**Not Found**.'
  schemas:
    nutzer:
      type: object
      properties:
        email:
          type: string
          description: '**E-Mail-Adresse** des Nutzers.'
        passwort:
          type: string
          description: '**Passwort** des Nutzers.'
        geschlecht:
          type: string
          description: '**Geschlecht** des Nutzers.'
        benutzername:
          type: string
          description: '**Benutzername** des Nutzers.'
        geburtsdatum:
          type: string
          format: date
          description: '**Geburtsdatum** des Users im Format "YYYY-MM-DD".'
      required:
        - email
        - benutzername
        - passwort
        - geburtsdatum
        - geschlecht
...
    kommentar:
      type: object
      properties:
        erstellungsdatum:
          type: string
          format: date-time
          description: '**Erstellungsdatum** des Kommentars im Format "YYYY-MM-DD HH:MM:SS".'
        text:
          type: string
          description: '**Text** des Kommentars.'
      required:
        - erstellungsdatum
        - text
...
    schlagwort:
      type: object
      properties:
        bezeichnung:
          type: string
          description: '**Bezeichnung** des Schlagwortes.'
      required:
        - bezeichnung
...